Make ESP32-vs1053_ext into a valid Arduino library
Since the library source files are placed in the src subfolder, the library is considered to be in Arduino Library 1.5 format. In order for libraries in this format to be recognized by the Arduino IDE, a library.properties file must be located in the root library folder. This file contains library metadata.

Reference: https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification